fix(#690): move SYSTEM_PATH_PREFIX from core to contracts/constants
SYSTEM_PATH_PREFIX was defined in core/nexus_fs_core.py but imported by
services (search_service, search_listing_mixin) creating a layering
violation. Moved the constant to contracts/constants.py where both core
and services can import it without cross-tier dependency.
