+++ src/nexus/services/search_listing_mixin.py
@@ -207,7 +207,7 @@ class SearchListingMixin:
                             logger.debug("Skipping deleted cross-zone path: %s", ct_path)
 
         # Filter out internal system entries
-        from nexus.core.nexus_fs_core import SYSTEM_PATH_PREFIX
+        from nexus.contracts.constants import SYSTEM_PATH_PREFIX
 
         all_files = [m for m in all_files if not m.path.startswith(SYSTEM_PATH_PREFIX)]
 
@@ -1003,7 +1003,7 @@ class SearchListingMixin:
                 zone_id=list_zone_id,
             )
 
-            from nexus.core.nexus_fs_core import SYSTEM_PATH_PREFIX
+            from nexus.contracts.constants import SYSTEM_PATH_PREFIX
 
             batch.items = [
                 item for item in batch.items if not item.path.startswith(SYSTEM_PATH_PREFIX)
